uta-analytics/vector/watcher/requirements.txt
watchdog==4.0.0
kafka-python-ng>=2.2.2
